added user tests
