+++ .travis.yml
@@ -1,0 +1,9 @@
+language: python
+python:
+    - "3.5"
+install:
+    - "pip install -r requirements.txt"
+    - "pip install -r test_requirements.txt"
+    - "pip install coveralls"
+script: "python setup.py test"
+after_success: "coveralls"
